Fixed issue where double as divided by int

+++ app/src/main/java/io/connorwyatt/flashcards/services/FlashcardTestService.kt
@@ -30,7 +30,7 @@ class FlashcardTestService {
                 }
             }
 
-            averageRating = total / flashcardTests.size
+            averageRating = total / flashcardTests.size.toDouble()
         }
 
         return averageRating
